Added a __main__.py
